make codebase cleaner
