+++ src/Router.tsx
@@ -2,11 +2,10 @@ import React from 'react';
 import { BrowserRouter as ReactRouter, Route, Switch } from 'react-router-dom';
 import Header from './components/layout/Header';
 import Footer from './components/layout/Footer';
-import MyParticles from './components/Particles';
-import Information from './components/Information';
-import BSCBridge from './components/BSCBridge';
-import NFT from './components/NFT/NFT';
-import NFTGallery from './components/NFT/NFTGallery';
+import HomePage from './pages/Home';
+import NFT from './pages/NFT/NFT';
+import NFTGallery from './pages/NFT/NFTGallery';
+import TeamPage from './pages/Team';
 import ComingSoon from './components/ComingSoon';
 
 const Router = () => {
@@ -14,20 +13,11 @@ const Router = () => {
     <ReactRouter>
       <Header />
       <Switch>
-        <Route
-          exact
-          path='/'
-          render={() => (
-            <>
-              <MyParticles />
-              <Information />
-            </>
-          )}
-        />
-        <Route exact path='/nft' render={() => <NFT />} />
-        <Route exact path='/bridge' render={() => <BSCBridge />} />
-        <Route exact path='/gallery' render={() => < NFTGallery/>} />
-        <Route exact path='/coming-soon' render={() => < ComingSoon/>} />
+        <Route exact path='/' component={HomePage} />
+        <Route exact path='/nft' component={NFT} />
+        <Route exact path='/gallery' component={NFTGallery} />
+        <Route exact path='/coming-soon' component={ComingSoon} />
+        <Route exact path='/team' component={TeamPage} />
       </Switch>
       <Footer />
     </ReactRouter>
