+++ src/components/layout/Footer.tsx
@@ -2,10 +2,21 @@
 import './Footer.css';
 
 import React from 'react';
+import { withRouter } from 'react-router-dom';
+import { History } from 'history';
 
-const Footer = () => {
+import BSCBridge from '../../components/BSCBridge';
+
+interface FooterProps {
+  history : History
+}
+
+const Footer:React.SFC<FooterProps>= ({ history }) => {
+  const backToHome = () => history.push('/')
+
   return (
     // <!-- Site footer -->
+    <>
     <footer className="site-footer">
       <div className="container">
         <div className="row">
@@ -47,7 +58,24 @@ const Footer = () => {
         </div>
       </div>
     </footer>
+    <div className="modal fade" id="exampleModalToggle" aria-hidden="true" aria-labelledby="exampleModalToggleLabel" >
+        <div className="modal-dialog modal-fullscreen">
+          <div className="modal-content">
+            <div className="modal-header">
+              <h5 className="modal-title" id="exampleModalToggleLabel"><img src="avalanche-logo.png" width="40" height="40" alt="" /> <img className="binancelogo" src="binance-logo.png" width="40" height="40" alt="" /> Spore Avalanche - BSC Bridge </h5>
+              <button type="button" className="close-modal" data-bs-dismiss="modal" aria-label="Close"><i className="fa fa-times"></i></button>
+            </div>
+            <div className="modal-body">
+              <BSCBridge />
+            </div>
+            <div className="modal-footer">
+              <button className="btn btn-primary claimair m-air" data-bs-dismiss="modal" onClick={backToHome}>Back to Home</button>
+            </div>
+          </div>
+        </div>
+      </div>
+    </>
   );
 }
 
-export default Footer;
+export default withRouter(Footer);
